packages/quantum-language-optimizer/quantum_language_optimizer-0.1.0.tar.gz/quantum_language_optimizer-0.1.0/quantum_language_optimizer/models/gpt_quantum.py
from .base_model import BaseModel
from ..core.quantum_rnn import QuantumRNN

class GPTQuantum(BaseModel):
    """
    GPTQuantum is a model that uses QuantumRNN for optimization.
    It extends the BaseModel class.
    """
    def __init__(self, num_qubits):
        """
        Initialize the GPTQuantum model.

        Parameters:
        num_qubits (int): The number of qubits to use in the QuantumRNN.
        """
        super().__init__()
        self.quantum_rnn = QuantumRNN(num_qubits)

    def train(self, data):
        """
        Train the GPTQuantum model with the provided data.

        Parameters:
        data (any): The data used for training the model.

        Note: This is a placeholder method. Implement the training logic as needed.
        """
        # Implement training logic here
        print("Training the GPTQuantum model...")

    def predict(self, input_sequence):
        """
        Predict the output for the given input sequence using QuantumRNN.

        Parameters:
        input_sequence (list of str): The input sequence data for which predictions are to be made.

        Returns:
        list of dict: The prediction results for each input in the sequence.
        """
        print("Predicting using the GPTQuantum model...")
        return self.quantum_rnn.forward(input_sequence)
