packages/quantum-language-optimizer/quantum_language_optimizer-0.1.0.tar.gz/quantum_language_optimizer-0.1.0/quantum_language_optimizer/tests/test_gpt_quantum.py
import unittest
from quantum_language_optimizer.models.gpt_quantum import GPTQuantum

class TestGPTQuantum(unittest.TestCase):
    def setUp(self):
        self.gpt_quantum = GPTQuantum(num_qubits=3)

    def test_train(self):
        train_data = ["example training data"]
        self.gpt_quantum.train(train_data)
        # Add assertions to verify the training process if applicable

    def test_predict(self):
        input_sequence = ['101', '110', '011']
        predictions = self.gpt_quantum.predict(input_sequence)
        self.assertEqual(len(predictions), 3)
        for result in predictions:
            self.assertIn('000', result)
            self.assertIn('111', result)

if __name__ == '__main__':
    unittest.main()
